local JS Match
A fix for Javascript files that are indeed local but may contain a get
request after .js .

Example can be find in the K2 component
components/com_k2/js/k2.js?v2.6.9&sitepath=/

Without the patch the javascript is no within the whole package but
remains alone. After the patch the script is embeddedin the scriptmerge
script url.